Visuals now working - Removed visualhelp
The error was that it was creating a 64bit pointer and my handler for
the point was 32bit, which caused loss of data. Working now and even
added a small animation to test.

+++ SimulationCPP/NaturalSelection/NaturalSelecton/GAUtils.cpp
@@ -69,7 +69,7 @@ float GAUtils::randFloatGen(){
 	std::mt19937 gen(rd());
 	std::uniform_int_distribution<> dist(1, 999);
 
-	return dist(gen) / 1000.0;
+	return (float)(dist(gen) / 1000.0);
 }
 
 /*
